Labeled Faces in the Wild
Labeled Faces in the Wild is a public benchmark for face verification, also known as pair matching. No matter what the performance of an algorithm on LFW, it should not be used to conclude that an algorithm is suitable for any commercial purpose. For more information please visit: http://vis-www.cs.umass.edu/lfw/#reference